building: use maven central or mary from environment
When publishing to maven central we don't want any other repositories listed in our pom.xml or (even if just compile-only) dependencies in it.

When you set the environment variable `MAVEN_REPOSITORY` to `CENTRAL` then it will comply with that. Otherwise (still the default) it will still use mary.

Question for the future: Should this be the default option? Maven central is pretty much the default and everybody is scared of custom repositories.